+++ test/cookbooks/test/recipes/config_4.rb
@@ -20,6 +20,7 @@ haproxy_listen 'admin' do
   ]
   http_response 'set-header Expires %[date(3600),http_date]'
   default_backend 'servers'
+  option ['dontlog-normal']
   extra_options('bind-process' => 'odd')
   hash_type 'consistent'
 end