+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This file is used to list changes made in each version of the haproxy cookbook.
 
 ## Unreleased
 
+- Add `option` property to `haproxy_listen` - [@eheydrick](https://github.com/eheydrick)
+
 ## 12.2.2 - *2021-10-05*
 
 ## 12.2.1 - *2021-08-30*